refactor(lab/17): start splitting some stuff up so it's easier to work on the collision thing
